+++ src/main/java/de/adorsys/keycloak/config/repository/IdentityProviderRepository.java
@@ -20,6 +20,7 @@
 
 package de.adorsys.keycloak.config.repository;
 
+import de.adorsys.keycloak.config.util.ResponseUtil;
 import org.keycloak.admin.client.resource.IdentityProviderResource;
 import org.keycloak.admin.client.resource.IdentityProvidersResource;
 import org.keycloak.representations.idm.IdentityProviderRepresentation;
@@ -27,6 +28,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.ws.rs.NotFoundException;
+import javax.ws.rs.core.Response;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -35,17 +38,14 @@ public class IdentityProviderRepository {
     private final RealmRepository realmRepository;
 
     @Autowired
-    public IdentityProviderRepository(
-            RealmRepository realmRepository
-    ) {
+    public IdentityProviderRepository(RealmRepository realmRepository) {
         this.realmRepository = realmRepository;
     }
 
-    public Optional<IdentityProviderRepresentation> tryToFindIdentityProvider(String realm, String name) {
+    public Optional<IdentityProviderRepresentation> tryToFindIdentityProvider(String realm, String alias) {
         Optional<IdentityProviderRepresentation> maybeIdentityProvider;
 
-        IdentityProvidersResource identityProvidersResource = realmRepository.loadRealm(realm).identityProviders();
-        IdentityProviderResource identityProviderResource = identityProvidersResource.get(name);
+        IdentityProviderResource identityProviderResource = loadIdentityProviderByAlias(realm, alias);
 
         try {
             maybeIdentityProvider = Optional.of(identityProviderResource.toRepresentation());
@@ -56,9 +56,22 @@ public class IdentityProviderRepository {
         return maybeIdentityProvider;
     }
 
+    public IdentityProviderRepresentation getIdentityProviderByAlias(String realm, String alias) {
+        IdentityProviderResource identityProviderResource = loadIdentityProviderByAlias(realm, alias);
+        if (identityProviderResource == null) {
+            return null;
+        }
+        return identityProviderResource.toRepresentation();
+    }
+
+    public List<IdentityProviderRepresentation> getIdentityProviders(String realm) {
+        return realmRepository.loadRealm(realm).identityProviders().findAll();
+    }
+
     public void createIdentityProvider(String realm, IdentityProviderRepresentation identityProvider) {
         IdentityProvidersResource identityProvidersResource = realmRepository.loadRealm(realm).identityProviders();
-        identityProvidersResource.create(identityProvider);
+        Response response = identityProvidersResource.create(identityProvider);
+        ResponseUtil.throwOnError(response);
     }
 
     public void updateIdentityProvider(String realm, IdentityProviderRepresentation identityProviderToUpdate) {
@@ -67,5 +80,18 @@ public class IdentityProviderRepository {
                 .get(identityProviderToUpdate.getAlias());
 
         identityProviderResource.update(identityProviderToUpdate);
+    }
+
+    public void deleteIdentityProvider(String realm, IdentityProviderRepresentation identityProviderToDelete) {
+        IdentityProviderResource identityProviderResource = realmRepository.loadRealm(realm)
+                .identityProviders()
+                .get(identityProviderToDelete.getInternalId());
+
+        identityProviderResource.remove();
+    }
+
+    private IdentityProviderResource loadIdentityProviderByAlias(String realm, String identityProviderAlias) {
+        return realmRepository.loadRealm(realm)
+                .identityProviders().get(identityProviderAlias);
     }
 }