+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Create, Update, Delete IdentityProviderMappers
+- Support for only updating changed IdentityProviders
+- Support for managed IdentityProviders
+
 ### Changed
 
 ### Fixed